Added every layout file for every screen size.
Note that the activity_node.xml files in every layout directory needs
an AD_UNIT_ID or else the activity will not run properly. The
AD_UNIT_ID is the AdMob Publisher ID. I am not committing this to the
public repository. I'm pretty sure there are no serious threats with
making the ID public, but it'd be confusing to commit this since this
is a unique per-user data point.

Signed-off-by: Byron Sanchez <bfs5030@gmail.com>