Updated layouts (activity_node.xml) to use new bookmark toggle icon.
The new bookmark has been switched from a regular image button to a
toggle button. The toggle button uses two different image resources.
The addition symbol is used to add a story to the bookmarks list, and
the subtraciton symbol is used to remove a story from the bookmarks
list.

Signed-off-by: Byron Sanchez <bfs5030@gmail.com>